Refactor GroupContainer for GroupStats and Contributors
