+++ packages/app-root/src/app/groups/[groupId]/contributors/page.js
@@ -1,0 +1,15 @@
+import GroupContributorsContainer from './GroupContributorsContainer'
+
+export const metadata = {
+  title: 'Group Contributors',
+  description: 'Zooniverse group contributors page'
+}
+
+export default function GroupContributors({ params, searchParams }) {
+  return (
+    <GroupContributorsContainer
+      groupId={params.groupId}
+      joinToken={searchParams.join_token}
+    />
+  )
+}
