+++ packages/lib-user/src/components/shared/GroupContainer/helpers/getUserGroupStatus.js
