+++ packages/lib-user/src/components/shared/GroupContainer/index.js
@@ -1,0 +1,1 @@
+export { default } from './GroupContainer'
